clicking outside components for showing/hiding
