+++ filter/src/App.js
@@ -1,18 +1,69 @@
-import React, { useState } from "react";
+import React, { useState, useRef, useEffect } from "react";
 import "./scss/filter.scss";
-import SideDrawer from "./components/SideDrawer";
+import RightSideDrawer from "./components/RightSideDrawer";
+import LeftSideDrawer from "./components/LeftSideDrawer";
+
+function useComponentVisible(initialIsVisible) {
+  const [showSideDrawer, setShowSideDrawer] = useState(false);
+  const ref = useRef(null);
+
+  const handleHideDropdown = (event: KeyboardEvent) => {
+    if (event.key === "Escape") {
+      setShowSideDrawer(false);
+    }
+  };
+
+  const handleClickOutside = (event) => {
+    if (ref.current && !ref.current.contains(event.target)) {
+      setShowSideDrawer(false);
+    }
+  };
+
+  useEffect(() => {
+    document.addEventListener("keydown", handleHideDropdown, true);
+    document.addEventListener("click", handleClickOutside, true);
+    return () => {
+      document.removeEventListener("keydown", handleHideDropdown, true);
+      document.removeEventListener("click", handleClickOutside, true);
+    };
+  });
+
+  return { ref, showSideDrawer, setShowSideDrawer };
+}
 
 function App() {
-  const [showSideDrawer, setShowSideDrawer] = useState(false);
+  const [labelName, setLabelName] = useState();
+  const [labelType, setLabelType] = useState();
+
+  const passValues = (filterName, filterType) => {
+    setLabelName(filterName);
+    setLabelType(filterType);
+  };
+
+  const { ref, showSideDrawer, setShowSideDrawer } = useComponentVisible(true);
+
   return (
-    <div className="App">
-      <input
-        type="submit"
-        value="+ add Filter"
-        onClick={() => setShowSideDrawer(true)}
-        className="dummy"
-      />
-      {showSideDrawer ? <SideDrawer /> : null}
+    <div ref={ref}>
+      {showSideDrawer && (
+        <div className="sideDrawer" ref={ref}>
+          <div className="row">
+            <div className="col-md-5 col-lg-5">
+              <LeftSideDrawer handleValue={passValues} />
+            </div>
+            <div className="col-md-7 col-lg-7">
+              <RightSideDrawer name={labelName} type={labelType} />
+            </div>
+          </div>
+        </div>
+      )}
+      <div>
+        <input
+          type="submit"
+          value="+ add filter"
+          className="dummy"
+          onClick={() => setShowSideDrawer(true)}
+        ></input>
+      </div>
     </div>
   );
 }
